# perf: improved func ValidateNodeForTxn node id parsing